backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.118.0
filelock==3.13.1
fonttools==4.60.1
fsspec==2024.6.1
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.12.0.88
packaging==25.0
pillow==11.0.0
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.1.0
pydantic==2.12.0
pydantic_core==2.41.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.13.3
torch==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.207
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
